Move common.geoserver to separate geoserver module
