+++ src/layman/geoserver_proxy_test.py
@@ -5,11 +5,11 @@ import requests
 from owslib.feature.schema import get_schema as get_wfs_schema
 import pytest
 
+from geoserver.util import get_layer_thumbnail, get_layer_square_bbox
 from layman import app, settings
 from layman.layer import db, util as layer_util
 from layman.layer.geoserver import wfs as geoserver_wfs
 from layman.layer.qgis import util as qgis_util, wms as qgis_wms
-from layman.common.geoserver import get_layer_thumbnail, get_layer_square_bbox
 
 
 @pytest.mark.usefixtures('ensure_layman')
