+++ src/layman/layer/geoserver/sld.py
@@ -1,6 +1,7 @@
-from layman.layer.filesystem import input_style
-from layman.common import geoserver
+from geoserver import util as gs_util
 from layman import settings, patch_mode
+from layman.common.db import launder_attribute_name
+from layman.layer.filesystem import input_style
 from . import wms
 from ...util import url_for
 
@@ -9,12 +10,12 @@ PATCH_MODE = patch_mode.DELETE_IF_DEPENDANT
 
 def get_workspace_style_url(workspace, style=None):
     geoserver_workspace = wms.get_geoserver_workspace(workspace)
-    return geoserver.get_workspace_style_url(geoserver_workspace, style)
+    return gs_util.get_workspace_style_url(geoserver_workspace, style)
 
 
 def get_workspace_layer_url(workspace, layer=None):
     geoserver_workspace = wms.get_geoserver_workspace(workspace)
-    return geoserver.get_workspace_layer_url(geoserver_workspace, layer)
+    return gs_util.get_workspace_layer_url(geoserver_workspace, layer)
 
 
 def pre_publication_action_check(workspace, layername):
@@ -31,7 +32,7 @@ def patch_layer(workspace, layername):
 
 def delete_layer(workspace, layername):
     geoserver_workspace = wms.get_geoserver_workspace(workspace)
-    sld_stream = geoserver.delete_workspace_style(geoserver_workspace, layername, auth=settings.LAYMAN_GS_AUTH)
+    sld_stream = gs_util.delete_workspace_style(geoserver_workspace, layername, auth=settings.LAYMAN_GS_AUTH)
     wms.clear_cache(workspace)
     if sld_stream:
         result = {
@@ -45,7 +46,7 @@ def delete_layer(workspace, layername):
 
 
 def get_layer_info(workspace, layername):
-    r = get_style_response(workspace, layername, geoserver.headers_sld, settings.LAYMAN_GS_AUTH)
+    r = get_style_response(workspace, layername, gs_util.headers_sld, settings.LAYMAN_GS_AUTH)
     if r.status_code == 200:
         url = url_for('rest_workspace_layer_style.get', workspace=workspace, layername=layername)
         info = {
@@ -67,7 +68,7 @@ def get_publication_uuid(workspace, publication_type, publication_name):
 def create_layer_style(workspace, layername):
     geoserver_workspace = wms.get_geoserver_workspace(workspace)
     style_file = input_style.get_layer_file(workspace, layername)
-    geoserver.post_workspace_sld_style(geoserver_workspace, layername, style_file)
+    gs_util.post_workspace_sld_style(geoserver_workspace, layername, style_file, launder_attribute_name)
     wms.clear_cache(workspace)
 
 
@@ -77,4 +78,4 @@ def get_metadata_comparison(workspace, layername):
 
 def get_style_response(workspace, stylename, headers=None, auth=None):
     geoserver_workspace = wms.get_geoserver_workspace(workspace)
-    return geoserver.get_workspace_style_response(geoserver_workspace, stylename, headers, auth)
+    return gs_util.get_workspace_style_response(geoserver_workspace, stylename, headers, auth)
