+++ src/layman/layer/geoserver/util.py
@@ -3,8 +3,8 @@ import requests
 from owslib.wms import WebMapService
 from owslib.wfs import WebFeatureService
 
+from geoserver.util import headers_json, get_proxy_base_url
 from layman.cache.mem import CACHE as MEM_CACHE
-from layman.common.geoserver import headers_json, get_proxy_base_url
 
 
 CACHE_GS_PROXY_BASE_URL_KEY = f'{__name__}:GS_PROXY_BASE_URL'
