+++ src/layman/layer/geoserver/wfs.py
@@ -2,10 +2,11 @@ from urllib.parse import urljoin
 import requests
 from flask import current_app
 
+from geoserver import util as gs_util
 from layman import settings, patch_mode
 from layman.cache import mem_redis
+from layman.common import geoserver as gs_common
 from layman.layer.util import is_layer_task_ready
-from layman.common import geoserver as common_geoserver
 from layman import util as layman_util
 from layman.layer import LAYER_TYPE
 from .util import get_gs_proxy_base_url
@@ -30,24 +31,24 @@ def post_layer(workspace, layername):
 
 
 def patch_layer(workspace, layername, title, description, access_rights=None):
-    common_geoserver.patch_feature_type(workspace, layername, title, description, settings.LAYMAN_GS_AUTH)
+    gs_util.patch_feature_type(workspace, layername, title, description, settings.LAYMAN_GS_AUTH)
     clear_cache(workspace)
 
     if access_rights and access_rights.get('read'):
-        security_read_roles = common_geoserver.layman_users_to_geoserver_roles(access_rights['read'])
-        common_geoserver.ensure_layer_security_roles(workspace, layername, security_read_roles, 'r', settings.LAYMAN_GS_AUTH)
+        security_read_roles = gs_common.layman_users_to_geoserver_roles(access_rights['read'])
+        gs_util.ensure_layer_security_roles(workspace, layername, security_read_roles, 'r', settings.LAYMAN_GS_AUTH)
 
     if access_rights and access_rights.get('write'):
-        security_write_roles = common_geoserver.layman_users_to_geoserver_roles(access_rights['write'])
-        common_geoserver.ensure_layer_security_roles(workspace, layername, security_write_roles, 'w', settings.LAYMAN_GS_AUTH)
+        security_write_roles = gs_common.layman_users_to_geoserver_roles(access_rights['write'])
+        gs_util.ensure_layer_security_roles(workspace, layername, security_write_roles, 'w', settings.LAYMAN_GS_AUTH)
 
 
 def delete_layer(workspace, layername):
-    common_geoserver.delete_feature_type(workspace, layername, settings.LAYMAN_GS_AUTH)
+    gs_util.delete_feature_type(workspace, layername, settings.LAYMAN_GS_AUTH)
     clear_cache(workspace)
 
-    common_geoserver.delete_security_roles(f"{workspace}.{layername}.r", settings.LAYMAN_GS_AUTH)
-    common_geoserver.delete_security_roles(f"{workspace}.{layername}.w", settings.LAYMAN_GS_AUTH)
+    gs_util.delete_security_roles(f"{workspace}.{layername}.r", settings.LAYMAN_GS_AUTH)
+    gs_util.delete_security_roles(f"{workspace}.{layername}.w", settings.LAYMAN_GS_AUTH)
     return {}
