+++ src/layman/layer/geoserver/wms.py
@@ -3,9 +3,10 @@ from urllib.parse import urlencode, urlparse, urlunparse, parse_qs, parse_qsl
 import requests
 from flask import current_app
 
+from geoserver import util as gs_util
 from layman import settings, patch_mode, util as layman_util
 from layman.cache import mem_redis
-from layman.common import geoserver as common_geoserver
+from layman.common import geoserver as gs_common
 from layman.layer.util import is_layer_task_ready
 from layman.layer import LAYER_TYPE, util as layer_util
 from .util import get_gs_proxy_base_url
@@ -33,27 +34,27 @@ def patch_layer(workspace, layername, title, description, access_rights=None):
     geoserver_workspace = get_geoserver_workspace(workspace)
     info = layer_util.get_layer_info(workspace, layername)
     if info['style_type'] == 'sld':
-        common_geoserver.patch_feature_type(geoserver_workspace, layername, title, description, settings.LAYMAN_GS_AUTH)
+        gs_util.patch_feature_type(geoserver_workspace, layername, title, description, settings.LAYMAN_GS_AUTH)
         clear_cache(workspace)
 
     if access_rights and access_rights.get('read'):
-        security_read_roles = common_geoserver.layman_users_to_geoserver_roles(access_rights['read'])
-        common_geoserver.ensure_layer_security_roles(geoserver_workspace, layername, security_read_roles, 'r', settings.LAYMAN_GS_AUTH)
+        security_read_roles = gs_common.layman_users_to_geoserver_roles(access_rights['read'])
+        gs_util.ensure_layer_security_roles(geoserver_workspace, layername, security_read_roles, 'r', settings.LAYMAN_GS_AUTH)
 
     if access_rights and access_rights.get('write'):
-        security_write_roles = common_geoserver.layman_users_to_geoserver_roles(access_rights['write'])
-        common_geoserver.ensure_layer_security_roles(geoserver_workspace, layername, security_write_roles, 'w', settings.LAYMAN_GS_AUTH)
+        security_write_roles = gs_common.layman_users_to_geoserver_roles(access_rights['write'])
+        gs_util.ensure_layer_security_roles(geoserver_workspace, layername, security_write_roles, 'w', settings.LAYMAN_GS_AUTH)
 
 
 def delete_layer(workspace, layername):
     geoserver_workspace = get_geoserver_workspace(workspace)
-    common_geoserver.delete_feature_type(geoserver_workspace, layername, settings.LAYMAN_GS_AUTH)
-    common_geoserver.delete_wms_layer(geoserver_workspace, layername, settings.LAYMAN_GS_AUTH)
-    common_geoserver.delete_wms_store(geoserver_workspace, settings.LAYMAN_GS_AUTH, get_qgis_store_name(layername))
+    gs_util.delete_feature_type(geoserver_workspace, layername, settings.LAYMAN_GS_AUTH)
+    gs_util.delete_wms_layer(geoserver_workspace, layername, settings.LAYMAN_GS_AUTH)
+    gs_util.delete_wms_store(geoserver_workspace, settings.LAYMAN_GS_AUTH, get_qgis_store_name(layername))
     clear_cache(workspace)
 
-    common_geoserver.delete_security_roles(f"{geoserver_workspace}.{layername}.r", settings.LAYMAN_GS_AUTH)
-    common_geoserver.delete_security_roles(f"{geoserver_workspace}.{layername}.w", settings.LAYMAN_GS_AUTH)
+    gs_util.delete_security_roles(f"{geoserver_workspace}.{layername}.r", settings.LAYMAN_GS_AUTH)
+    gs_util.delete_security_roles(f"{geoserver_workspace}.{layername}.w", settings.LAYMAN_GS_AUTH)
     return {}
