+++ src/layman/upgrade/upgrade_v1_9.py
@@ -1,6 +1,7 @@
 import logging
 
 from db import util as db_util
+from geoserver import util as gs_util
 from layman import settings
 from layman.common import geoserver as gs_common
 
@@ -62,7 +63,7 @@ where w.name = %s
             security_roles = gs_common.layman_users_to_geoserver_roles(users_roles)
             logger.info(f'    Setting security roles for: ({workspace}/{publication_name}).{right_type} '
                         f'to ({security_roles}) from layman roles ({users_roles})')
-            gs_common.ensure_layer_security_roles(workspace, publication_name, security_roles, right_type[0], settings.LAYMAN_GS_AUTH)
+            gs_util.ensure_layer_security_roles(workspace, publication_name, security_roles, right_type[0], settings.LAYMAN_GS_AUTH)
 
     logger.info(f'    DONE - access rights EVERYONE is not propagated to GeoServer for authenticated users')
 
@@ -76,10 +77,10 @@ def geoserver_remove_users_for_public_workspaces():
     auth = settings.LAYMAN_GS_AUTH
     for (workspace, ) in public_workspaces:
         logger.info(f'      Delete user and role for workspace {workspace}')
-        role = gs_common.username_to_rolename(workspace)
-        gs_common.delete_user_role(workspace, role, auth)
-        gs_common.delete_user_role(workspace, settings.LAYMAN_GS_ROLE, auth)
-        gs_common.delete_role(role, auth)
-        gs_common.delete_user(workspace, auth)
+        role = gs_util.username_to_rolename(workspace)
+        gs_util.delete_user_role(workspace, role, auth)
+        gs_util.delete_user_role(workspace, settings.LAYMAN_GS_ROLE, auth)
+        gs_util.delete_role(role, auth)
+        gs_util.delete_user(workspace, auth)
 
     logger.info(f'    DONE - delete unnecessary users and roles created for public workspaces')
